[feat] Server로부터 데이터를 받아서 group listing하는 로직 구현
- getGroupInfo API를 만들고 그 안에서 서버로부터 데이터를 받은 뒤
state에 setting하는 형식으로 로직 구현
- GroupScreen 안에 있던 GroupListScreen을 다른 파일로 분리
